refactor : Distribute Application to MVCS
Distribute Application's logic to controller, model, service, view.
