+++ src/main/java/racingcar/controller/RacingController.java
@@ -1,0 +1,4 @@
+package racingcar.controller;
+
+public class RacingController {
+}
