+++ src/main/java/racingcar/model/Racing.java
@@ -1,0 +1,4 @@
+package racingcar.model;
+
+public class Racing {
+}
